Detect aliveness of node links more robustly
`Weak::ptr_eq(&weak, &Weak::new())` returns true only when `weak`
is the value created by `Weak::new()`.
If the weak reference was dereferenceable at some time and after that
it is invalidated, comparing to `&Weak::new()` by means of
`Weak::ptr_eq` will return false.

The weak reference here is not expected to be invalidated during
referred from other nodes (because someone will make it alive until
it become orphan), but use more robust comparison for more
future-proofness (in case of the weak links will be allowed to be
invalidated in future).