Modify membership correctly when moving nodes to another tree
About copying membership
------------------------

Before structure edit locking is implemented, memberships are just
refcounted references to the core of the tree the node belongs to.
So, it is completely safe and valid to clone one node's membership
and use it for other nodes in the same tree.

However, after structure edit locking is implemented, `MembershipCore`s
have node-specific data (namely the aggregated lock counts through
the node) and it is now invalid to use memberships for non-corresponding
nodes.

By changes in this commit, now `Node`s are created from the membership
corresponding to the node of interest.

About replacing tree core of a node link
----------------------------------------

A node link (and its membership core data) is associated to the tree
and the lock through a node is also associated to the same tree.
When the associated tree is switched, the lock state through a node
must be preserved in a new tree, so the lock need to be transferred.